+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+### Added
+- Added `craft\base\ElementInterface::getCanonicalUid()`. ([#9757](https://github.com/craftcms/cms/issues/9757))
+
 ### Changed
 - Edit Entry pages now show a “Save draft” button for users without the “Publish live changes” permission. ([#9750](https://github.com/craftcms/cms/issues/9750))
 - Users without the “Delete entries” permission can now delete their own unpublished drafts.